Changed intensity to double

+++ Models/PetriParameters.cs
@@ -4,8 +4,8 @@ public class PetriParameters
 {
     public int MaxQueueLength { get; init; }
     public int NumOfProcessingUnits { get; init; }
-    public int ClientsIntensity { get; init; }
-    public int ProcessingIntensity { get; init; }
+    public double ClientsIntensity { get; init; }
+    public double ProcessingIntensity { get; init; }
     public int ImitationTime { get; init; }
     public bool DisplayAllProcesses { get; init; }
 }
